palette.c the variables red, green, blue are declared as int but cfg_get_color expects guint16 types. This could lead to potential issues on systems where the size of int is larger than guint16. Changing their types to guint16 ensures compatibility with the cfg_get_color function.

+++ src/fe-gtk3/palette.c
@@ -127,7 +127,7 @@ palette_load (void)
 	char prefname[256];
 	struct stat st;
 	char *cfg;
-	int red, green, blue;
+	guint16 red, green, blue;
 	int upgrade = FALSE;
 
 	fh = xchat_open_file ("colors.conf", O_RDONLY, 0, 0);
